finetune-llm-ontology/requirements.txt
accelerate
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.43.0
certifi
charset-normalizer
click
datasets==2.18.0
dill==0.3.8
docstring_parser==0.16
frozenlist==1.4.1
fsspec
gmpy2
huggingface-hub==0.21.4
idna
Jinja2
joblib
markdown-it-py==3.0.0
MarkupSafe
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
mpmath
multidict==6.0.5
multiprocess==0.70.16
networkx
numpy
packaging
pandas==2.2.1
peft
protobuf==3.20.3
psutil
pyarrow==15.0.2
pyarrow-hotfix==0.6
pydantic
pydantic_core
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv
pytz==2024.1
PyYAML
regex
requests
rich==13.7.1
sacremoses==0.0.43
safetensors
scikit-learn
scipy
shtab==1.7.1
six
sympy
threadpoolctl
tokenizers==0.15.2
torch==2.2.1
tqdm
transformers==4.39.1
triton==2.2.0
trl==0.8.1
typing_extensions
tyro==0.7.3
tzdata==2024.1
urllib3
xxhash==3.4.1
yarl==1.9.4
